adjust llvm and related packages
